Refactor to an idiomatic Writable stream
Previously, the Interpreter was implemented as a Node.js Transform
stream in order to utilize the `_flush` method specific to that class.
This was done in order to perform an asynchronous operation once the
stream was closing--an operation that deferred the subsequent "finish"
event. However, this implementation detail influenced the API in ways
that conflict with traditional stream usage in Node.js.

By over-riding the built-in `Writable.prototype.end` method, the desired
behavior can be achieved while promoting usage patterns that are more
in-line with typical Node.js streams.